requirements.txt
# Job Performance Monitor - Python Dependencies

# HTTP client
requests>=2.28.0

# Databricks SDK for Python
databricks-sdk>=0.20.0

# Environment file support
python-dotenv>=1.0.0

# Web UI server
flask>=3.0.0
flask-cors>=4.0.0
